+++ src/models/quantized_qwen3_asr/shaders/embedding_q8_0_f32.py
@@ -45,7 +45,7 @@ EMBEDDING_Q8_0_F32 = ShaderVariant(
                 name='output',
                 io_kind=IOKind.OUTPUT,
                 role='output',
-                contract=TensorContract(dtype='float32', shape=('I0', 'I1', 'H',)),
+                contract=TensorContract(dtype='float16', shape=('I0', 'I1', 'H',)),
             ),
         ),
         push_constants=PushConstantSpec(
@@ -63,12 +63,12 @@ EMBEDDING_Q8_0_F32 = ShaderVariant(
 #version 450
 #extension GL_EXT_shader_explicit_arithmetic_types_int16 : require
 #extension GL_EXT_shader_16bit_storage : require
-
+#extension GL_EXT_shader_explicit_arithmetic_types_float16 : require
 #extension GL_EXT_shader_explicit_arithmetic_types_int64 : require
 layout(std430) buffer;
 layout(set = 0, binding = 0) buffer restrict readonly WeightBuffer { uint16_t weight[]; };
 layout(set = 0, binding = 1) buffer restrict readonly IndicesBuffer { int64_t indices[]; };
-layout(set = 0, binding = 2) buffer restrict writeonly OutputBuffer { float output_values[]; };
+layout(set = 0, binding = 2) buffer restrict writeonly OutputBuffer { float16_t output_values[]; };
 layout(push_constant) uniform PushConstants { uint num_indices; uint embedding_dim; } pc;
 layout(local_size_x = 256, local_size_y = 1, local_size_z = 1) in;
 
@@ -93,7 +93,7 @@ void main() {
     const uint token_idx = idx / pc.embedding_dim;
     const uint dim_idx = idx - token_idx * pc.embedding_dim;
     const int64_t token_id = indices[token_idx];
-    output_values[idx] = q8_0_value(uint(token_id), dim_idx);
+    output_values[idx] = float16_t(q8_0_value(uint(token_id), dim_idx));
 }
 """,
 )
